Enable WiFi
It's possible that not all of these workarounds are necessary, but the fact that there's impurity (as discussed in the links in comments) makes testing every configuration painful. In the end I had to reboot to make things take effect.

I thought for a while that it still wasn't working. But actually it's just that my methodology for checking that WiFi was enabled was faulty - neither `ip a` nor `rfkill list` showed anything to suggest it was working. They do now, though. This is probably related to the above, fixed, issue.

It can take a while (average maybe half an hour) to be able to even ping the WiFi address after a reboot, or a `nixos-rebuild switch` that affects anything relevant. Again, hopefully the same issue.