Stop running most systemd services as root
For one thing, it should silence some Dhall XDG errors in `tennis-scraper`. Probably better for security as well.

Ideally we'd find some way to forward port 80 to another, so that we needn't run Droopy as root. But neither `virtualisation.forwardPorts` or `networking.nat.forwardPorts` seem to do that sort of thing.