Recover from `curl` failure in `email-ip`
I _think_ this only fails when the device has no wi-fi connection, including potentially in the first few seconds after a boot. But I can also now properly monitor this, and maybe log the exit code (in cases I've seen so far, always `6/NOTCONFIGURED`) if we need to debug further.