Remove root command pipe
It was only when choosing whether to add a HTTP endpoint for it that I figured it's a potential security flaw.

We've never actually had any real use case for it. In ed365cc we hypothesised using it to reboot, which I think was prompted by my having some serious issues SSHing in to the system at some point. But even then, we could do that easily enough through a bespoke system service.

If we do ever wish to use the removed functionality, e.g. for a permanent reboot button, we can just go back to supporting a particular finite set of commands.