Simplify `uinput` group creation
I guess I was wrongly assuming that adding to `extraGroups` was enough to trigger group creation. And groups like `input` and `wheel` in some sense already exist.

Note that these sorts of changes seem to require a reboot, which added to my confusion when experimenting with this. I suspect it would just be a logout if not for my "lingering" hack.