Don't use error LED outside Haskell script
Systemd generally doesn't really seem to like `postStop` running forever. For one thing, it makes `systemctl restart` not work, which is a pretty major issue.

It's a shame. It seemed elegant in a way. Otherwise to get error LED working we'd need a full system for tracking which processes are causing LED to be on. Then again, centralising that would make it easier to see which process is responsible for the LED being on, although this should be obvious from emails, unless having email troubles.

But we have email working pretty reliably, so it's not the end of the world.

It's also somewhat cleaner to go back to only one process having ownership of the LED.