Add service for monitoring web pages
Note that it's just easier (maybe only possible in practice) to manipulate with jq than dhall, especially once we want to iterate over list containing structured data.

Also, `-t` in `readarray` is to remove trailing newlines. It may not usually make a difference here but it's safer in general.

We use a `while` loop rather than `systemd` restarts so that we can control the pause time via the Dhall config.