Check that pipes are actually pipes
I have seen an issue, whose exact cause I'm unsure of, where one of the other processes creates the file before the activation script has fired, and thus the activation script doesn't replace it with an actual pipe. This eventually causes all sorts of trouble.

We could go further and emit a warning (via stdout or even email) when this does happen, but I think this will do for now.