Use much lower debounce period
We choose not to worry about spurious "falling" events from pressing (as opposed to releasing) the button. Testing now with `gpiomon --edges falling --chip 0 23`, they appear to be less than a one-in-ten event. This is better than I remember seeing previously, which could be to do with having recently given the wires a firm shove in to place... Anwyay, our hand has actually been somewhat forced by an apparent bug in `libgpiod`'s new debouncing support (until recently we rolled our own): it would appear\* that debouncing is totally broken, at least with large periods, since when both the press and release trigger a "falling" event, neither is reported.

In the long run, this should be fixed by better wiring. For now, we don't actually use the GPIO button for anything critical or everyday anyway.

\* Though this is quite difficult to observe. We'd ideally monitor the same button both with and without debouncing enabled, but this isn't possible as each process grabs a lock on the pin.