Add Mandelbrot set as wallpaper
We could add `picture-uri` as well as `picture-uri-dark` but this seems unnecessary given that we stick to the dark theme. Setting a wallpaper through the Gnome settings UI does set, along with `org/gnome/desktop/screensaver.picture-uri`, which is also irrelevant in this setup.

The PNG was generated from https://github.com/georgefst/hs-scripts/ with:
```hs
width = 3840
height = 3840
(xMin, xMax) = (-3, 1.8)
(yMin, yMax) = (-2.4, 2.4)
```

I considered actually adding the relevant Haskell code as a dependency, but it doesn't seem worth requiring a 15-second build (the image generation is very unoptimised) whenever there's a minor GHC version bump or similar. I may revisit this idea though. It would be cool to have different machines show different interesting parts of the set, for example.