Use super key and space to close windows
Much less fiddly than alt-f4, particularly on most modern keyboards without separate fn and media keys.

We add an extra alt for switching input source, which previously used this binding. The backward version was automatically updated to match when using the Gnome settings UI, so we may as well be consistent with that.