Add Crow
This is my personal desktop, an iMac Pro which triple-boots with MacOS and Windows, and previously ran EndeavourOS.

Config is for now essentially just Fry without the Obsidian stuff, and with some changes for Mac hardware.

It's a bit unfortunate that we're using a different Nixpkgs to the other machines, but the T2 Linux NixOS install process defaulted to unstable, and so we have `system.stateVersion = "25.11"`, and it doesn't seem safe therefore to move backwards. We'll get everything on to 25.11 when that comes out. The requirement for users to be in a group (`group = "gthomas"`) is presumably a new post-25.05 requirement.

The install also added the new cache to `trusted-substituters` as well as `substituters` but this doesn't seem to have any real effect for our workflow.

We needed `users.mutableUsers = false;` in order for the user to appear in the Gnome login screen. While this must be a bug, and the Reddit comment which suggested it found that it could actually be reverted after it had done it's job of kicking some imperative state somewhere, it seems simpler, particularly in terms of re-installs, to just keep it around. I don't really want mutable users anyway.

We'll revisit the module and folder structure soon, as it's all starting to seem a bit clunky. In particular, the repetition in `flake.nix`, and the use of `common-users.nix` due to the Obsidian module setting up its own users.

Some minor stuff has been cleaned up slightly in passing, e.g. consistent comments.