Enable fractional scaling
Non-Wayland apps do get worse, but Spotify is the only one I use much, and even there [support is on the way](https://community.spotify.com/t5/Desktop-Linux/Wayland-support/td-p/5231525/highlight/true/page/6) (it can be made to work with a flag, but there's an ugly border on Gnome at least due to CSD issues).

We've set 150% for the scaling on Crow. Note that this settings is not stored in `dconf`, but in `~/.config/monitors.xml`, which is not currently tracked by this Nix config.