Configure Crow keyboard layout
It's possible that it'd be better to enable both layouts on both devices, since we might want to use the Mac keyboard with Fry, or indeed a "normal" keyboard with Crow. We even have a keybinding for changing the input source, and it's explicitly configured since the original one overlapped with what we wanted for closing windows. Still, the clutter in the tray (top right) is mildly annoying for something we'd very rarely use. It wouldn't be a surprise though if we later revisit this decision.