Avoid fixing per-workspace tiling layouts
Locking this actually just doesn't really work once we start adding and removing workspaces, with the extension seemingly getting confused about why it can't change the layout.

This more dynamic approach is in line with how we're launching windows anyway. We're not declaring a single set of windows in particular places which will always be open, but rather focusing on making it easy to put them wherever we want, even though we will often be setting up the same configuration in practice.

For now, there's no support for setting these layouts with the keyboard, but support is coming with the next extension release. And using the mouse/trackpad isn't too much of an issue as it's not something we'll actually change all that often, and the extension remembers the configuration to some extent.