Add another split layout for large monitors
This has been fine-tuned for my 3840x2560 monitor, at 150% scaling, to perfectly fit two 120-width editors in VSCode with all panels closed. The y value was chosen in order to be as small as possible to fit the Gather V2 pop-out window in the corner. One minor annoyance is that when the Gnome terminal is used to fill the remaining space, it's slightly too small to show the tab titles. We may revisit this when Gnome supports more of a range of fractions.