Set keybinding for toggling "always on top"
Super+A is chosen for consistency with my old XFCE configs.

This conflicts with the default keybinding for the app view, which I barely use (preferring to just use Super and start typing, rather than browse), and prefer Super+0 for anyway, as an xtension of using Super+numkeys for launching dock apps, which I use constantly.