Add Gnome extension for window scripting
We were initially alerted to its existence due to `gnomeExtensions.window-calls-extended`, but actually this one is all we need.

This was used in conjunction with `ydotool` for a script for sending F5 commands to Firefox. It allows us to list windows, and to activate with `gdbus call --session --dest org.gnome.Shell --object-path /org/gnome/Shell/Extensions/Windows --method org.gnome.Shell.Extensions.Windows.Activate $id`.