Add keybinding for copying current line from terminal
Ideally, it'd be great to be able to navigate output of long-running commands without using the mouse as well, but realistically that would probably mean ditching Gnome's terminal which I otherwise like, as this bug has been open for 23 years: https://gitlab.gnome.org/GNOME/vte/-/issues/588.

This is the only sensible keybinding which seemed to work, as any combination with ctrl-c killed commands, and the common ctrl-p conflicts with VSCode.