Bump all flake inputs
The commit used for pinning the old kernel is the one immediately before 6.16 was removed, not that we really need to be quite that precise. The problem essentially is that ZFS is best used with LTS kernels, as it's often the case that there's no other non-EOL version which it supports, and thus the version can go up and down. But the latest LTS kernel doesn't work with our hardware, so we don't have that option. See here for much more detail: https://discourse.nixos.org/t/aggressive-kernel-removal-on-eol-in-nixos/23097/75.