Move Fry and Clark to Nixpkgs unstable
It's simpler to have both desktops on the same Nixpkgs, especially for overlays-based changes we'd like to explore. All devices will be moving to 25.11 soon anyway.

For now we fix the pin we were already using for unstable. Using the latest commit breaks our Gnome extensions as Nixpkgs bumped to Gnome 49 and extensions support is about two weeks behind.