Revert "Remove VSCode shell level workaround"
This reverts commit ec486e7584048c39c217e9c62eb0d24350dbf78d.

On second thoughts, there's no real relevant difference between VSCode and, say, the Gnome terminal. It's weird seeing the dot for standard Bash shells. And when we want a `nix develop` shell, it'll usually be the default for the project, so it's fine not to see a dot. That'd just be like if `nix-env-selector` were a bit smarter and wrapped the shells for us anyway.