Add keybinding for switching tiling layouts
This is something we've wanted for a while, and has just recently been added. Unfortunately, there isn't yet an option to cycle backwards: https://github.com/domferr/tilingshell/issues/424.

Monitors are now the wrong way round in the menu since the update, but this is probably a basic transient bug, and not really an issue now that we use the keyboard for switching instead. It is a bit unfortunate though that the switcher shows on the primary screen rather than the one in question.